cimapp/requirements.txt
astroid==1.6.1
backports.functools-lru-cache==1.5
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
dj-static==0.0.6
django-suit==0.2.25
enum34==1.1.6
futures==3.2.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.8.2
pyodbc==4.0.22
pytz==2018.3
pywin32==222
requests==2.18.4
screen==1.0.1
singledispatch==3.4.0.3
six==1.11.0
static3==0.7.0
urllib3==1.22
virtualenv==15.1.0
wrapt==1.10.11
xlwt==1.3.0